@@ -1,8 +1,8 @@
 module github.com/flostadler/name-generator
 
-go 1.13
+go 1.18
 
 require (
-	github.com/docker/docker v1.13.1
-	github.com/gorilla/mux v1.7.1
+	github.com/docker/docker v20.10.17+incompatible
+	github.com/gorilla/mux v1.8.0
 )
